Load MSDataSetGenerator via IServiceProvider
The previous code relied on the class being in the GAC, which is not
guaranteed for generators, although it was true for this one.  The code
was also more complex and relied on knowledge of the layout of the
Visual Studio registry contents.

Instead, use the IServiceProvider interface of the DTE to load the
class.  This is both simpler and should be more robust going forward.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>